requirements-loose.txt
# Core requirements with compatibility focus for Python 3.13
# Flask ecosystem with specific versions for compatibility
# Flask==2.3.2 - installed separately
# flask-restx==1.2.0 - installed separately
# Werkzeug==2.3.7 - installed separately
# jsonschema==4.17.3 - installed separately

# Other dependencies with minimum versions
alembic>=1.13.0
aniso8601>=9.0.0
beautifulsoup4>=4.12.0
blinker>=1.7.0
bs4>=0.0.1
click>=8.1.0
croniter>=2.0.0
Flask-Admin>=1.6.0
Flask-Mail>=0.9.1
Flask-Migrate>=4.0.0
Flask-RQ2>=18.3
Flask-Script>=2.0.6
Flask-SQLAlchemy>=3.0.0
isodate>=0.6.0
itsdangerous>=2.1.0
Jinja2>=3.1.0
Mako>=1.3.0
MarkupSafe>=2.1.0
psycopg2-binary>=2.9.9
python-dateutil>=2.8.0
python-editor>=1.0.0
pytz>=2024.1
redis>=5.0.0
rq>=1.15.0
rq-scheduler>=0.10.0
six>=1.16.0
SQLAlchemy>=2.0.0
tzlocal>=5.0.0
WTForms>=2.1
